+++ types/task/schema/index.ts
@@ -1,0 +1,15 @@
+import { TaskStateEnum } from '..';
+import { User } from '../..';
+
+export type BaseType = {
+  _id: string;
+  label: string;
+  owner: User['_id'];
+  state: TaskStateEnum;
+
+  createAt: Date;
+  updateAt: Date;
+};
+
+export * from './task';
+export * from './task-group';
