+++ types/task/schema/task-group.ts
@@ -1,0 +1,9 @@
+import { BaseType, Task } from '.';
+
+export type TaskGroup = BaseType & {
+  current: Task['_id'][];
+  order: number;
+  processing: number;
+  wait: number;
+  total: number;
+};
