+++ types/task/schema/task.ts
@@ -1,0 +1,8 @@
+import { BaseType } from '.';
+import { ImageFileSchema, ImageThumbnailFileSchema } from '../..';
+
+export type Task = BaseType & {
+  group?: string;
+  order: number;
+  file?: ImageFileSchema['_id'] | ImageThumbnailFileSchema['_id'];
+};
